fix: thread actual model name through embedding storage
Add `modelName(): string` to the `AIProvider` interface and all
implementations (NullProvider → "null", OllamaProvider → embedModel,
GeminiProvider → embedModel). Extract the duplicated
`generateEpisodeEmbeddings` helper from git.ts and markdown.ts into a
shared `src/ai/utils.ts` that calls `provider.modelName()` instead of
the hardcoded sentinel "provider". Fix NullProvider.embed to return one
empty vector per input text (was returning []). Document the
findSimilar 100-embedding cap as a known v0.1 limitation.
