+++ packages/engram-core/src/ingest/git.ts
@@ -10,11 +10,11 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 import { ulid } from "ulid";
 import type { AIProvider } from "../ai/provider.js";
+import { generateEpisodeEmbeddings } from "../ai/utils.js";
 import type { EngramGraph } from "../format/index.js";
 import { ENGINE_VERSION } from "../format/version.js";
 import { resolveEntity } from "../graph/aliases.js";
 import { addEdge } from "../graph/edges.js";
-import { storeEmbedding } from "../graph/embeddings.js";
 import { addEntity, type EvidenceInput } from "../graph/entities.js";
 import { addEpisode } from "../graph/episodes.js";
 import { supersedeEdge } from "../temporal/supersession.js";
@@ -736,64 +736,5 @@ export async function ingestGitRepo(
     const msg = err instanceof Error ? err.message : String(err);
     failIngestionRun(graph, runId, msg);
     throw err;
-  }
-}
-
-/**
- * Generate embeddings for a batch of episode IDs using the given provider.
- * Never throws — embedding failures are logged and skipped.
- */
-async function generateEpisodeEmbeddings(
-  graph: EngramGraph,
-  provider: AIProvider,
-  episodeIds: string[],
-): Promise<void> {
-  if (episodeIds.length === 0) return;
-
-  interface EpisodeContentRow {
-    id: string;
-    content: string;
-  }
-
-  // Fetch episode content
-  const rows: EpisodeContentRow[] = [];
-  for (const id of episodeIds) {
-    const row = graph.db
-      .query<EpisodeContentRow, [string]>(
-        "SELECT id, content FROM episodes WHERE id = ?",
-      )
-      .get(id);
-    if (row) rows.push(row);
-  }
-
-  if (rows.length === 0) return;
-
-  try {
-    const texts = rows.map((r) => r.content);
-    const embeddings = await provider.embed(texts);
-
-    for (let i = 0; i < rows.length; i++) {
-      const embedding = embeddings[i];
-      if (!embedding || embedding.length === 0) continue;
-
-      try {
-        storeEmbedding(
-          graph,
-          rows[i].id,
-          "episode",
-          "provider",
-          embedding,
-          rows[i].content.slice(0, 500),
-        );
-      } catch (err: unknown) {
-        const msg = err instanceof Error ? err.message : String(err);
-        console.warn(
-          `[engram] generateEpisodeEmbeddings: skip ${rows[i].id}: ${msg}`,
-        );
-      }
-    }
-  } catch (err: unknown) {
-    const msg = err instanceof Error ? err.message : String(err);
-    console.warn(`[engram] generateEpisodeEmbeddings: provider error: ${msg}`);
   }
 }
