+++ packages/engram-core/src/ingest/markdown.ts
@@ -10,9 +10,9 @@ import { createHash } from "node:crypto";
 import * as fs from "node:fs";
 import * as path from "node:path";
 import type { AIProvider } from "../ai/provider.js";
+import { generateEpisodeEmbeddings } from "../ai/utils.js";
 import type { EngramGraph } from "../format/index.js";
 import { ENGINE_VERSION } from "../format/version.js";
-import { storeEmbedding } from "../graph/embeddings.js";
 import { addEpisode } from "../graph/episodes.js";
 import type { IngestResult } from "./git.js";
 
@@ -154,60 +154,4 @@ export async function ingestMarkdown(
   }
 
   return counts;
-}
-
-/**
- * Generate embeddings for a list of episode IDs.
- * Never throws — failures are logged and skipped.
- */
-async function generateEpisodeEmbeddings(
-  graph: EngramGraph,
-  provider: AIProvider,
-  episodeIds: string[],
-): Promise<void> {
-  interface EpisodeContentRow {
-    id: string;
-    content: string;
-  }
-
-  const rows: EpisodeContentRow[] = [];
-  for (const id of episodeIds) {
-    const row = graph.db
-      .query<EpisodeContentRow, [string]>(
-        "SELECT id, content FROM episodes WHERE id = ?",
-      )
-      .get(id);
-    if (row) rows.push(row);
-  }
-
-  if (rows.length === 0) return;
-
-  try {
-    const texts = rows.map((r) => r.content);
-    const embeddings = await provider.embed(texts);
-
-    for (let i = 0; i < rows.length; i++) {
-      const embedding = embeddings[i];
-      if (!embedding || embedding.length === 0) continue;
-
-      try {
-        storeEmbedding(
-          graph,
-          rows[i].id,
-          "episode",
-          "provider",
-          embedding,
-          rows[i].content.slice(0, 500),
-        );
-      } catch (err: unknown) {
-        const msg = err instanceof Error ? err.message : String(err);
-        console.warn(
-          `[engram] generateEpisodeEmbeddings: skip ${rows[i].id}: ${msg}`,
-        );
-      }
-    }
-  } catch (err: unknown) {
-    const msg = err instanceof Error ? err.message : String(err);
-    console.warn(`[engram] ingestMarkdown: provider error: ${msg}`);
-  }
 }
