+++ packages/engram-core/test/ai/null.test.ts
@@ -10,14 +10,18 @@ import { NullProvider } from "../../src/ai/null.js";
 describe("NullProvider", () => {
   const provider = new NullProvider();
 
+  test("modelName() returns 'null'", () => {
+    expect(provider.modelName()).toBe("null");
+  });
+
   test("embed() returns empty array for empty input", async () => {
     const result = await provider.embed([]);
     expect(result).toEqual([]);
   });
 
-  test("embed() returns empty array even with text input", async () => {
+  test("embed() returns one empty vector per input text", async () => {
     const result = await provider.embed(["hello world", "foo bar"]);
-    expect(result).toEqual([]);
+    expect(result).toEqual([[], []]);
   });
 
   test("extractEntities() returns empty array", async () => {
@@ -31,7 +35,7 @@ describe("NullProvider", () => {
   });
 
   test("embed() never throws", async () => {
-    await expect(provider.embed(["text"])).resolves.toEqual([]);
+    await expect(provider.embed(["text"])).resolves.toEqual([[]]);
   });
 
   test("extractEntities() never throws", async () => {
